refactor: formatter 적용
